refactor(fast_io): mirror iocp module tree in iocp_stub (SPL-30 #2458)
Split the monolithic 1087-LoC `iocp_stub.rs` into a directory whose
submodule layout mirrors `crates/fast_io/src/iocp/`. Each stub item now
lives in the file matching its real Windows counterpart: config, error,
disk_batch, file_reader, file_writer, file_factory, pump, socket. Tests
move into a dedicated `tests` submodule.

Zero API change: the module is still exposed as `crate::iocp` via the
`#[path = "iocp_stub/mod.rs"]` attribute on the cross-platform fallback,
and `mod.rs` re-exports the same public surface as before.
